Moves exceptions to package
