+++ domain/src/main/java/de/cashflow/domain/exception/UserExistsException.kt
@@ -1,4 +1,4 @@
-package de.cashflow.domain.repository
+package de.cashflow.domain.exception
 
 class UserExistsException(p0: String?) : RuntimeException(p0) {
 }
