+++ domain/src/main/java/de/cashflow/domain/exception/UserNotFoundException.kt
@@ -1,4 +1,4 @@
-package de.cashflow.domain.repository
+package de.cashflow.domain.exception
 
 class UserNotFoundException(p0: String?) : Exception(p0) {
 }
